+++ packages/help-center/src/components/help-center-chat.tsx
@@ -8,7 +8,6 @@ import { useEffect } from '@wordpress/element';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useHelpCenterContext } from '../contexts/HelpCenterContext';
 import { useChatStatus, useShouldUseWapuu } from '../hooks';
-import { ExtraContactOptions } from './help-center-extra-contact-option';
 import './help-center-chat.scss';
 
 export function HelpCenterChat( {
@@ -54,9 +53,6 @@ export function HelpCenterChat( {
 			userFieldFlowName={ userFieldFlowName ?? params.get( 'userFieldFlowName' ) }
 			isUserEligibleForPaidSupport={ isUserEligibleForPaidSupport }
 			forceEmailSupport={ Boolean( forceEmailSupport ) }
-			extraContactOptions={
-				<ExtraContactOptions isUserEligible={ isUserEligibleForPaidSupport } />
-			}
 		>
 			<div className="help-center__container-chat">
 				<OdieAssistant />
