--- packages/help-center/src/components/help-center-extra-contact-option.scss
@@ -1,6 +1,0 @@
-.help-center__container-extra-contact-options {
-	padding: 0;
-	margin: 0;
-	border: none;
-	min-height: 68px;
-}
