+++ packages/odie-client/src/context/index.tsx
@@ -72,7 +72,6 @@ export const OdieAssistantProvider: React.FC< OdieAssistantProviderProps > = ( {
 	isUserEligibleForPaidSupport = true,
 	canConnectToZendesk = false,
 	isLoadingCanConnectToZendesk = false,
-	extraContactOptions,
 	selectedSiteId,
 	selectedSiteURL,
 	userFieldMessage,
@@ -199,7 +198,6 @@ export const OdieAssistantProvider: React.FC< OdieAssistantProviderProps > = ( {
 				setChat: setMainChatState,
 				clearChat,
 				currentUser,
-				extraContactOptions,
 				isChatLoaded,
 				isMinimized,
 				experimentVariationName,
